+++ src/main/java/vn/fpt/diamond_shop/request/CreateDetailJewelryRequest.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class CreateDetailJewelryRequest {
 
-    Integer idJewelry;
+    Long idJewelry;
 
 
 }
