+++ src/main/java/vn/fpt/diamond_shop/request/GetDetailJewelryRequest.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class GetDetailJewelryRequest {
 
-    Integer idJewelry;
+    Long idJewelry;
 
 
 }
